# cmd/relay/openapi.yaml
openapi: 3.1.0
info:
  title: Relay Webhook Delivery API
  description: |
    Relay is a webhook delivery system that guarantees delivery through
    idempotency keys, persistent storage, exponential backoff retries,
    and circuit breakers.

    ## Authentication
    All API requests require an API key passed in the `X-API-Key` header.

    ## Rate Limiting
    API requests are rate limited per API key. Check response headers for limits:
    - `X-RateLimit-Limit`: Maximum requests per window
    - `X-RateLimit-Remaining`: Remaining requests in current window
    - `X-RateLimit-Reset`: Unix timestamp when the window resets
  version: 1.0.0
  contact:
    name: Relay Support
    url: https://github.com/relay/relay
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.relay.example.com
    description: Production
...
paths:
...
  /api/v1/alerting/rules:
    get:
      tags:
        - Alerting
      summary: List alert rules
      description: Returns all configured alert rules
      operationId: listAlertRules
      responses:
        "200":
          description: List of alert rules
          content:
            application/json:
              schema:
                type: object
                properties:
                  rules:
                    type: array
                    items:
                      $ref: "#/components/schemas/AlertRule"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      tags:
        - Alerting
      summary: Create an alert rule
      description: Creates a new alert rule for monitoring webhook delivery
      operationId: createAlertRule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAlertRuleRequest"
      responses:
        "201":
          description: Alert rule created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertRule"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/alerting/rules/{ruleId}:
    get:
      tags:
        - Alerting
      summary: Get alert rule by ID
      description: Returns a single alert rule
      operationId: getAlertRule
      parameters:
        - $ref: "#/components/parameters/RuleId"
      responses:
        "200":
          description: Alert rule details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertRule"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      tags:
        - Alerting
      summary: Update an alert rule
      description: Updates an existing alert rule
      operationId: updateAlertRule
      parameters:
        - $ref: "#/components/parameters/RuleId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAlertRuleRequest"
      responses:
        "200":
          description: Alert rule updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertRule"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags:
        - Alerting
      summary: Delete an alert rule
      description: Deletes an alert rule
      operationId: deleteAlertRule
      parameters:
        - $ref: "#/components/parameters/RuleId"
      responses:
        "204":
          description: Alert rule deleted successfully
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/alerting/rules/{ruleId}/enable:
    post:
      tags:
        - Alerting
      summary: Enable an alert rule
      description: Enables a previously disabled alert rule
      operationId: enableAlertRule
      parameters:
        - $ref: "#/components/parameters/RuleId"
      responses:
        "200":
          description: Alert rule enabled successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertRule"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/alerting/rules/{ruleId}/disable:
    post:
      tags:
        - Alerting
      summary: Disable an alert rule
      description: Disables an alert rule without deleting it
      operationId: disableAlertRule
      parameters:
        - $ref: "#/components/parameters/RuleId"
      responses:
        "200":
          description: Alert rule disabled successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertRule"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/alerting/evaluate:
    post:
      tags:
        - Alerting
      summary: Evaluate alert rules
      description: Manually triggers evaluation of all enabled alert rules
      operationId: evaluateAlertRules
      responses:
        "200":
          description: Evaluation completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  evaluated:
                    type: integer
                    description: Number of rules evaluated
                  triggered:
                    type: integer
                    description: Number of alerts triggered
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/alerting/history:
    get:
      tags:
        - Alerting
      summary: Get alert history
      description: Returns the history of triggered alerts
      operationId: getAlertHistory
      parameters:
        - name: limit
          in: query
          description: Maximum number of alerts to return
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
      responses:
        "200":
          description: Alert history
          content:
            application/json:
              schema:
                type: object
                properties:
                  alerts:
                    type: array
                    items:
                      $ref: "#/components/schemas/Alert"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # Connector Endpoints
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
  parameters:
...
    RuleId:
      name: ruleId
      in: path
      required: true
      description: Alert rule ID
      schema:
        type: string
...
  schemas:
...
    AlertRule:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the alert rule
        name:
          type: string
          description: Human-readable name for the rule
        description:
          type: string
          description: Optional description of what the rule monitors
        enabled:
          type: boolean
          description: Whether the rule is actively being evaluated
        condition:
          $ref: "#/components/schemas/AlertCondition"
        actions:
          type: array
          items:
            $ref: "#/components/schemas/AlertAction"
          description: Actions to take when the alert fires
        cooldownMinutes:
          type: integer
          description: Minimum minutes between repeated alerts for the same rule
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - enabled
        - condition
        - actions
        - createdAt
        - updatedAt
    AlertCondition:
      type: object
      properties:
        metric:
          $ref: "#/components/schemas/AlertMetric"
        operator:
          $ref: "#/components/schemas/AlertOperator"
        threshold:
          type: number
          description: Value to compare the metric against
        windowMinutes:
          type: integer
          description: Time window in minutes for metric evaluation
        endpointId:
          type: string
          format: uuid
          nullable: true
          description: Optional endpoint ID to scope the metric to
      required:
        - metric
        - operator
        - threshold
        - windowMinutes
    AlertMetric:
      type: string
      enum:
        - failure_rate
        - dead_letter_count
        - latency_p99
        - queue_depth
        - circuit_breaker_trips
      description: |
        Metric to monitor:
        - failure_rate: Percentage of failed deliveries
        - dead_letter_count: Number of events moved to dead letter queue
        - latency_p99: 99th percentile delivery latency in milliseconds
        - queue_depth: Number of events waiting in queue
        - circuit_breaker_trips: Number of circuit breaker trips
    AlertOperator:
      type: string
      enum:
        - gt
        - gte
        - lt
        - lte
        - eq
      description: |
        Comparison operator:
        - gt: Greater than
        - gte: Greater than or equal
        - lt: Less than
        - lte: Less than or equal
        - eq: Equal to
    AlertAction:
      type: object
      properties:
        type:
          $ref: "#/components/schemas/AlertActionType"
        config:
          type: object
          description: Configuration for the action (varies by type)
      required:
        - type
        - config
    AlertActionType:
      type: string
      enum:
        - slack
        - email
        - webhook
        - pagerduty
      description: |
        Type of action to perform:
        - slack: Send Slack notification
        - email: Send email notification
        - webhook: Call a webhook URL
        - pagerduty: Create PagerDuty incident
    Alert:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the alert instance
        ruleId:
          type: string
          description: ID of the rule that triggered this alert
        ruleName:
          type: string
          description: Name of the rule that triggered this alert
        metric:
          $ref: "#/components/schemas/AlertMetric"
        value:
          type: number
          description: Actual metric value that triggered the alert
        threshold:
          type: number
          description: Threshold that was exceeded
        message:
          type: string
          description: Human-readable alert message
        triggeredAt:
          type: string
          format: date-time
          description: When the alert was triggered
      required:
        - id
        - ruleId
        - ruleName
        - metric
        - value
        - threshold
        - message
        - triggeredAt
    CreateAlertRuleRequest:
      type: object
      properties:
        name:
          type: string
          description: Human-readable name for the rule
          minLength: 1
          maxLength: 255
        description:
          type: string
          description: Optional description of what the rule monitors
        enabled:
          type: boolean
          default: true
        condition:
          $ref: "#/components/schemas/AlertCondition"
        actions:
          type: array
          items:
            $ref: "#/components/schemas/AlertAction"
          minItems: 1
        cooldownMinutes:
          type: integer
          minimum: 1
          default: 5
      required:
        - name
        - condition
        - actions
    UpdateAlertRuleRequest:
      type: object
      properties:
        name:
          type: string
          description: Human-readable name for the rule
        description:
          type: string
          description: Optional description of what the rule monitors
        enabled:
          type: boolean
        condition:
          $ref: "#/components/schemas/AlertCondition"
        actions:
          type: array
          items:
            $ref: "#/components/schemas/AlertAction"
        cooldownMinutes:
          type: integer
          minimum: 1

    # Connector Schemas
...
    Error:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        details:
          type: object
      required:
        - error
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "Invalid request body"
            code: "BAD_REQUEST"
    Unauthorized:
      description: Unauthorized - missing or invalid API key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "Invalid API key"
            code: "UNAUTHORIZED"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "Event not found"
            code: "NOT_FOUND"
...
security:
  - ApiKeyAuth: []